requirements-dev.txt
-r requirements.txt
django-debug-toolbar==4.4.6
flake8==7.1.1
isort==5.13.2
pylint==3.3.2
pylint-django==2.6.1
